add hijack mechanism to allow booting from stock kernel
when the stock kernel boots, it attempts to run /system/etc/init.qcom*
scripts. this commit sets up that script to be hijacked in such a way
that if we are not booted into kexec, then the kexec-boot zip is placed
on the cache partiiton before rebooting into recovery and auto-flashing
it. in this way even if we are booted from the stock kernel, we will
end up booted under kexec.